% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.R
\name{calculate_custom_score}
\alias{calculate_custom_score}
\title{Calculate the score of sentences}
\usage{
calculate_custom_score(text, positivewords, negativewords, check = 0)
}
\arguments{
\item{text}{A vector of sentences or a sentence (English).}

\item{positivewords}{A vector of words of positive sentiment.}

\item{negativewords}{A vector of words of negative sentiment.}

\item{check}{A flag variable whose value if 1 denotes appending of passed positive and negative words with existing words.By
default, this method will work with supplied positive and negative words only.}
}
\value{
A vector containing polarity of each sentence.
}
\description{
This function loads text and calculates score of each sentence on basis
of presence of words of positive and negative sentiment supplied externally as paramater, 
presence of negation,and checking for sarcasm. 0 indicates neutral sentiment. Positive value indicates
positive sentiment. Negative value indicates negative sentiment. 99 indicates
sarcasm.
}
\examples{
calculate_custom_score("This is good",c("good"),c("bad"),1)
calculate_custom_score(c("This is good","This is bad"),c("good"),c("bad"),0)
}

