% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.R
\name{calculate_custom_sentiment}
\alias{calculate_custom_sentiment}
\title{Predicts the sentiment of sentences}
\usage{
calculate_custom_sentiment(text, positivewords, negativewords, check = 0)
}
\arguments{
\item{text}{A vector of sentences or a sentence (English).}

\item{positivewords}{A vector of words of positive sentiment.}

\item{negativewords}{A vector of words of negative sentiment.}

\item{check}{A flag variable whose value if 1 denotes appending of passed positive and negative words with existing words.By
default, this method will work with supplied positive and negative words only.}
}
\value{
A vector containing sentiment of each sentence.
}
\description{
This function loads text and words of positive and negative sentiment supplied externally as paramater and calculates sentiment of each sentence. It classifies
sentences into 6 categories: Positive, Negative, Very Positive, Very Negative
Sarcasm and Neutral.
}
\examples{
calculate_custom_sentiment("This is good",c("good"),c("bad"),1)
calculate_custom_sentiment(c("This is good","This is bad"),c("good"),c("bad"),0)
}

