% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{calculate_score}
\alias{calculate_score}
\title{Calculate the score of sentences}
\usage{
calculate_score(text)
}
\arguments{
\item{text}{A vector of sentences or a sentence (English).}
}
\value{
A vector containing polarity of each sentence.
}
\description{
This function loads text and calculates score of each sentence on basis
of presence of words of positive and negative sentiment, presence of negation,
and checking for sarcasm. 0 indicates neutral sentiment. Positive value indicates
positive sentiment. Negative value indicates negative sentiment. 99 indicates
sarcasm.
}
\examples{
calculate_score("This is good")
calculate_score(c("This is good","This is bad"))
}

