% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{calculate_sentiment}
\alias{calculate_sentiment}
\title{Predicts the sentiment of sentences}
\usage{
calculate_sentiment(text)
}
\arguments{
\item{text}{A vector of sentences or a sentence (English).}
}
\value{
A vector containing sentiment of each sentence.
}
\description{
This function loads text and calculates sentiment of each sentence. It classifies
sentences into 6 categories: Positive, Negative, Very Positive, Very Negative
Sarcasm and Neutral.
}
\examples{
calculate_sentiment("This is good")
calculate_sentiment(c("This is good","This is bad"))
}

