% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.R
\name{calculate_custom_total_presence_sentiment}
\alias{calculate_custom_total_presence_sentiment}
\title{Calculate the number of sentences in each category of sentiment.}
\usage{
calculate_custom_total_presence_sentiment(text, positivewords, negativewords)
}
\arguments{
\item{text}{A vector of sentences or a sentence (English).}

\item{positivewords}{A vector of words of positive sentiment.}

\item{negativewords}{A vector of words of negative sentiment.}
}
\value{
A 2-D matrix with two rows and 6 columns where first row contains the name of sentiment
category and the second row contains the number in each category in string format.
}
\description{
This function loads text and words of positive and negative sentiment supplied externally as paramater, and calculates number of sentences which are positive,
negative, very positive, very negative, neutral and sarcasm.
}
\examples{
calculate_custom_total_presence_sentiment(c("This is good","This is bad"),c("good"),c("bad"))
}

